+++ test/object-store/realm.cpp
@@ -1359,6 +1359,148 @@ TEST_CASE("SharedRealm: convert") {
         REQUIRE(local_realm2->read_group().get_table("class_object")->size() == 1);
     }
 }
+
+TEST_CASE("SharedRealm: convert - embedded objects") {
+    TestSyncManager tsm;
+    ObjectSchema object_schema = {"object",
+                                  {
+                                      {"_id", PropertyType::Int, Property::IsPrimary{true}},
+                                      {"value", PropertyType::Int},
+                                      {"embedded_link", PropertyType::Object | PropertyType::Nullable, "embedded"},
+                                  }};
+    ObjectSchema embedded_schema = {"embedded",
+                                    ObjectSchema::ObjectType::Embedded,
+                                    {
+                                        {"name", PropertyType::String | PropertyType::Nullable},
+                                    }};
+    Schema schema{object_schema, embedded_schema};
+
+    SyncTestFile sync_config1(tsm.app(), "default");
+    sync_config1.schema = schema;
+    TestFile local_config1;
+    local_config1.schema = schema;
+    local_config1.schema_version = sync_config1.schema_version;
+
+    SECTION("can copy a synced realm to a synced realm") {
+        auto sync_realm1 = Realm::get_shared_realm(sync_config1);
+        sync_realm1->begin_transaction();
+
+        SECTION("null embedded object") {
+            sync_realm1->read_group().get_table("class_object")->create_object_with_primary_key(0);
+        }
+
+        SECTION("embedded object") {
+            auto obj = sync_realm1->read_group().get_table("class_object")->create_object_with_primary_key(0);
+            auto col_key = sync_realm1->read_group().get_table("class_object")->get_column_key("embedded_link");
+            obj.create_and_set_linked_object(col_key);
+        }
+
+        sync_realm1->commit_transaction();
+        wait_for_upload(*sync_realm1);
+        wait_for_download(*sync_realm1);
+
+        // Copy to a new sync config
+        SyncTestFile sync_config2(tsm.app(), "default");
+        sync_config2.schema = schema;
+
+        sync_realm1->convert(sync_config2);
+
+        auto sync_realm2 = Realm::get_shared_realm(sync_config2);
+
+        // Check that the data also exists in the new realm
+        REQUIRE(sync_realm2->read_group().get_table("class_object")->size() == 1);
+
+        // Verify that sync works and objects created in the new copy will get
+        // synchronized to the old copy
+        sync_realm2->begin_transaction();
+        sync_realm2->read_group().get_table("class_object")->create_object_with_primary_key(1);
+        sync_realm2->commit_transaction();
+        wait_for_upload(*sync_realm2);
+        wait_for_download(*sync_realm1);
+
+        sync_realm1->refresh();
+        REQUIRE(sync_realm1->read_group().get_table("class_object")->size() == 2);
+    }
+
+    SECTION("can convert a synced realm to a local realm") {
+        auto sync_realm = Realm::get_shared_realm(sync_config1);
+        sync_realm->begin_transaction();
+
+        SECTION("null embedded object") {
+            sync_realm->read_group().get_table("class_object")->create_object_with_primary_key(0);
+        }
+
+        SECTION("embedded object") {
+            auto obj = sync_realm->read_group().get_table("class_object")->create_object_with_primary_key(0);
+            auto col_key = sync_realm->read_group().get_table("class_object")->get_column_key("embedded_link");
+            obj.create_and_set_linked_object(col_key);
+        }
+
+        sync_realm->commit_transaction();
+        wait_for_upload(*sync_realm);
+        wait_for_download(*sync_realm);
+
+        sync_realm->convert(local_config1);
+
+        auto local_realm = Realm::get_shared_realm(local_config1);
+
+        // Check that the data also exists in the new realm
+        REQUIRE(local_realm->read_group().get_table("class_object")->size() == 1);
+    }
+
+    SECTION("can convert a local realm to a synced realm") {
+        auto local_realm = Realm::get_shared_realm(local_config1);
+        local_realm->begin_transaction();
+
+        SECTION("null embedded object") {
+            local_realm->read_group().get_table("class_object")->create_object_with_primary_key(0);
+        }
+
+        SECTION("embedded object") {
+            auto obj = local_realm->read_group().get_table("class_object")->create_object_with_primary_key(0);
+            auto col_key = local_realm->read_group().get_table("class_object")->get_column_key("embedded_link");
+            obj.create_and_set_linked_object(col_key);
+        }
+
+        local_realm->commit_transaction();
+
+        // Copy to a new sync config
+        local_realm->convert(sync_config1);
+
+        auto sync_realm = Realm::get_shared_realm(sync_config1);
+
+        // Check that the data also exists in the new realm
+        REQUIRE(sync_realm->read_group().get_table("class_object")->size() == 1);
+    }
+
+    SECTION("can copy a local realm to a local realm") {
+        auto local_realm1 = Realm::get_shared_realm(local_config1);
+        local_realm1->begin_transaction();
+
+        SECTION("null embedded object") {
+            local_realm1->read_group().get_table("class_object")->create_object_with_primary_key(0);
+        }
+
+        SECTION("embedded object") {
+            auto obj = local_realm1->read_group().get_table("class_object")->create_object_with_primary_key(0);
+            auto col_key = local_realm1->read_group().get_table("class_object")->get_column_key("embedded_link");
+            obj.create_and_set_linked_object(col_key);
+        }
+
+        local_realm1->commit_transaction();
+
+        // Copy to a new local config
+        TestFile local_config2;
+        local_config2.schema = schema;
+        local_config2.schema_version = local_config1.schema_version;
+        local_realm1->convert(local_config2);
+
+        auto local_realm2 = Realm::get_shared_realm(local_config2);
+
+        // Check that the data also exists in the new realm
+        REQUIRE(local_realm2->read_group().get_table("class_object")->size() == 1);
+    }
+}
 #endif // REALM_ENABLE_SYNC
 
 TEST_CASE("SharedRealm: async writes") {